Move ValidateDirectory functions to FileFolders for plugin usage
